+++ msamp/nn/__init__.py
@@ -7,6 +7,7 @@ from msamp.nn.parameter import ScalingParameter
 from msamp.nn.module import ScalingModule
 from msamp.nn.state import model_state
 from msamp.nn.linear import FP8Linear, LinearReplacer
+from msamp.nn import functional
 from msamp.nn.clip_grad import clip_grad_norm_
 
 __all__ = ['ScalingParameter', 'ScalingModule', 'model_state', 'FP8Linear', 'LinearReplacer', 'clip_grad_norm_']
